Do not load persistence file at gateway init
**Breaking change**

* Break out threading gateway class, ThreadingGateway. Change order of
  keyword arguments to clarify and simplify class inheritence.
* Move persistence file loading, saving and scheduling of saving to own
  method on ThreadingGateway. This allows future use of asyncio. The
  new method is called start_persistence. This must be called before
  starting the thread loop of the gateway, if persistence is used, to
  ensure persistence file is loaded before communication starts. This
  also allows the user control over when persistence should be loaded.
